Add first version of loopfs (not integrated with container yet).

+++ src/container.h
@@ -2,7 +2,7 @@ typedef struct container {
 	char *image_name;
 	char *init_script;
 	char *id;
-	int pid;
+	long pid;
 } container;
 
 container *start_container(char *image_name, char *init_script);
